# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BookMeArtist API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: number;
  artistId: number;
  artistName: string;
  artistImage?: string;
  clientName: string;
  clientEmail: string;
  eventDate: string;
  eventType: string;
  packageName?: string;
  budget: number;
  brief: string;
  location: string;
  status: BookingStatus;
  createdAt: Date;
}