# lib/api-zod/src/generated/types/createBookingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BookMeArtist API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingRequest {
  artistId: number;
  clientName: string;
  clientEmail: string;
  eventDate: string;
  eventType: string;
  packageName?: string;
  budget: number;
  brief: string;
  location: string;
}